+++ src/vanta.clouds.js
@@ -13,6 +13,7 @@ Effect.prototype.defaultOptions = {
   sunlightColor: 0xff9933, // 0x1a9eaa
   scale: 3,
   scaleMobile: 12,
+  speed: 1,
   mouseEase: true,
 }
 
@@ -22,6 +23,7 @@ uniform vec2 u_mouse;
 uniform float u_time;
 uniform sampler2D u_tex;
 
+uniform float speed;
 uniform vec3 skyColor;
 uniform vec3 cloudColor;
 uniform vec3 cloudShadowColor;
@@ -52,10 +54,10 @@ float noise( vec3 x ){
                    mix( iqhash(n+170.0), iqhash(n+171.0),f.x),f.y),f.z);
 }
 
-vec3 speed = vec3(0.5,0.01,1.0) * 0.5;
+vec3 speed1 = vec3(0.5,0.01,1.0) * 0.5 * speed;
 float constantTime = 1000.;
 float map5( in vec3 p ){
-    vec3 q = p - speed*(u_time + constantTime);
+    vec3 q = p - speed1*(u_time + constantTime);
     float f;
     f  = 0.50000*noise( q ); q = q*2.02;
     f += 0.25000*noise( q ); q = q*2.03;
@@ -65,7 +67,7 @@ float map5( in vec3 p ){
     return clamp( 1.5 - p.y - 2.0 + 1.75*f, 0.0, 1.0 );
 }
 float map4( in vec3 p ){
-    vec3 q = p - speed*(u_time + constantTime);
+    vec3 q = p - speed1*(u_time + constantTime);
     float f;
     f  = 0.50000*noise( q ); q = q*2.02;
     f += 0.25000*noise( q ); q = q*2.03;
@@ -74,7 +76,7 @@ float map4( in vec3 p ){
     return clamp( 1.5 - p.y - 2.0 + 1.75*f, 0.0, 1.0 );
 }
 float map3( in vec3 p ){
-    vec3 q = p - speed*(u_time + constantTime);
+    vec3 q = p - speed1*(u_time + constantTime);
     float f;
     f  = 0.50000*noise( q ); q = q*2.02;
     f += 0.25000*noise( q ); q = q*2.03;
@@ -82,7 +84,7 @@ float map3( in vec3 p ){
     return clamp( 1.5 - p.y - 2.0 + 1.75*f, 0.0, 1.0 );
 }
 float map2( in vec3 p ){
-    vec3 q = p - speed*(u_time + constantTime);
+    vec3 q = p - speed1*(u_time + constantTime);
     float f;
     f  = 0.50000*noise( q ); q = q*2.02;
     f += 0.25000*noise( q );
@@ -146,7 +148,7 @@ void main(){
     vec2 p = (-u_resolution.xy + 2.0*gl_FragCoord.xy)/ u_resolution.y;
 
     vec2 m = u_mouse.xy/u_resolution.xy;
-    m.y = (1.0 - m.y) * 0.3 + 0.25; // camera height
+    m.y = (1.0 - m.y) * 0.33 + 0.28; // camera height
 
     m.x *= 0.25;
     m.x += sin(u_time * 0.1 + 3.1415) * 0.25 + 0.25;